client-api: Rename r0::uiaa::authorize_fallback to get_uiaa_fallback_page
